### 2025-08-29 #8893 @editorialbot
[PRE REVIEW]: ArrowSpace: A primer on Spectral Indexing for vector search
**Submitting author:** @Mec-iS (<a href="http://orcid.org/0000-0002-8804-2963">Lorenzo Moriondo</a>)
**Repository:** https://github.com/Mec-iS/arrowspace-rs
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.0
