### 2025-08-29 #8893 @danielskatz
👋 @Mec-iS - thanks for your submission.

I notice your paper is about 4000 words long, while JOSS recommends 250-1000 words. You either need to dramatically reduce the content in your paper (perhaps pointing to content in the repo or docs), or submit to another venue that supports longer papers.

In addition, you could work on the possibly missing DOI that editorialbot suggests, but note that it may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right or you've changed the .md file's content to make a new PDF. editorialbot commands need to be the first entry in a new comment. You can also use the command `@editorialbot check repository` to check a bunch of things in the repository, including the word count of the paper.

Please let me know how you want to proceed.
