### 2025-08-29 24c3c57
reduce paper wording
