+++ paper.md
@@ -48,46 +48,6 @@ The treatment is particularly valuable for understanding how spectral properties
 Basic points:
 - Definition: for a feature row $x$ and item-Laplacian $L$, the smoothness is $E = \frac{x^\top L x}{x^\top x}$, which is non‑negative, scale‑invariant in $x$, near‑zero for constants on connected graphs, and larger for high‑frequency signals; the Rayleigh quotient is the normalised Dirichlet Energy, it is the discrete Dirichlet energy normalised by signal power.
 - Physical Interpretation: Dirichlet energy measure the "potential energy" or "stiffness" of a configuration while the Rayleigh quotient normalises this by the total "mass" or "signal power". the result is a scale-invariant measure of how much energy is required per unit mass (in our case the items-nodes).
-- The numerator equals the sum of weighted edge differences $\sum_{(i,j)} w_{ij}(x_i-x_j)^2$, directly capturing roughness over the graph, a classical link between Laplacians and Dirichlet energy used throughout spectral methods.
-
-Some implementation starting points:
-- Rayleigh energy $x^\top L x / x^\top x$ measures how "wiggly" a feature signal is over an item graph; constants yield near-zero on connected graphs, while alternating patterns are larger, making it a principled spectral smoothness score for search and structure discovery.
-- Pure Rayleigh can collapse near zero or be hard to compare across datasets; mapping energy to a bounded score and blending with a dispersion statistic produces a stable, comparable score that preserves spectral meaning while improving robustness for ranking and filtering.
-
-
-### Graph and data model
-Rayleigh energy score is complemented for spectral indexing by computing the graph Laplacian [@Spielman:2007] of the dataset:
-- Items and features: `ArrowSpace` stores a matrix with rows = feature signals and columns = items; the item graph nodes are the columns, and Rayleigh is evaluated per feature row against that item-Laplacian, aligning spectral scores with dataset geometry.
-- Item Laplacian: a Laplacian matrix is constructed over the graph of the items using a `λ`‑proximity policy (`ε` threshold on per‑item `λ`, union-symmetrised, k‑capped, kernel-weighted); diagonals store degrees and off‑diagonals are $−weights$, satisfying standard Laplacian invariants used by the Rayleigh quotient.
-
-Example:
-```rust
-// 1. Build item graph based on lambda proximity
-let items = vec![
-    vec![1.0, 2.0],  // Item 0: λ_0 = 0.3
-    vec![1.1, 2.1],  // Item 1: λ_1 = 0.35  
-    vec![3.0, 1.0],  // Item 2: λ_2 = 0.8
-];
-let aspace = ArrowSpace::from_items(...)
-
-// 2. Connect items with |λ_i - λ_j| ≤ ε
-// Items 0,1 connected (|0.3-0.35| = 0.05 ≤ ε)
-// Items 0,2 not connected (|0.3-0.8| = 0.5 > ε)
-// Items 1,2 not connected (|0.35-0.8| = 0.45 > ε)
-
-// 3. Resulting Laplacian (simplified):
-//     [  w   -w    0 ]
-// L = [ -w    w    0 ]  where w = kernel weight
-//     [  0    0    0 ]
-```
-
-### Role of Laplacian
-What the graph Laplacian contributes to Rayleigh energy:
-1. Spectral Smoothness: Captures how features vary across item relationships
-2. Graph Structure: Encodes similarity topology beyond simple pairwise distances
[…]
