### 2025-08-30 #8893 @Mec-iS
@danielskatz If it is OK I would like to resubmit a shorter version of ~1300 words with a modified title. Would that work? I can consider modifying down to less that 1000 words if that is a hard ceiling. Thank you.
