### 2025-09-02 e4e66ff
refactor bibliography
