+++ paper.bib
@@ -14,13 +14,6 @@
     Note = {Available at: \url{https://people.cs.umass.edu/~mahadeva/cs791bb/lectures-s2006/lec4.pdf}}
 }
 
-@misc{Grindrod:2020,
-    Author = {{Grindrod}, Peter},
-    Title = {{Laplacians of Complex Networks}},
-    Institution = {University of Bristol},
-    Note = {Available at: \url{https://people.maths.bris.ac.uk/~maajg/teaching/complexnets/laplacians.pdf}}
-}
-
 @misc{Chen:2020,
     Author = {{Chen}, Guangliang},
     Title = {{Math 253: Rayleigh Quotient Lecture Notes}},
@@ -72,13 +65,6 @@ z
     Doi = {10.1007/s11222-007-9033-z}
 }
 
-@misc{Chua:2025,
-    Author = {{Chua}, Dexter},
-    Title = {{Heat Kernel and Spectral Dimension}},
-    Note = {Available at: \url{https://dec41.user.srcf.net/exp/heat_kernel/heat_kernel.pdf}},
-    Year = 2025
-}
-
 @phdthesis{Bai:2007,
     Author = {{Bai}, Xiao},
     Title = {{Heat Kernel Analysis On Graphs}},
@@ -107,11 +93,4 @@ z
     Publisher = {University of Pennsylvania},
     Year = 2015,
     Note = {Available at: \url{https://www.physics.upenn.edu/biophys/PMLS/}}
-}
-
-@misc{Wikipedia:DirichletEnergy,
-    Title = {{Dirichlet energy}},
-    Publisher = {Wikipedia},
-    Note = {Available at: \url{https://en.wikipedia.org/wiki/Dirichlet_energy}},
-    Year = 2024
 }
