### 2025-09-02 #8893 @Mec-iS
OK. I modified the `paper.md`, please rerun the checks on the same repository.
