### 2025-09-02 #8893 @danielskatz
As I wrote above (https://github.com/openjournals/joss-reviews/issues/8893#issuecomment-3237168090), you can run the checks yourself.
