### 2025-09-02 #8893 @Mec-iS
> also, please let me know a bit about the development process you've gone through. Given the fact that there is a 4-day commit history, I'm unclear if this meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS.

Sure, let me expand on this. This research is the by-product of my years-long development at [smartcore](https://github.com/smartcorelib) (a numerical library for ML) where I am lead developer and code-owner. It has spun off from my current research on efficient Agentic AI workflows when tackling the problem of efficient vector retrieval, references:
- This piece of code was intended to be an `smartcore` implementation but took its own stage, probably it will be reimplemented using `smartcore` to be made production-ready
- Previous work carried on: W3C Web Agents WC [mailing list](https://www.w3.org/Search/Mail/Public/search?from=Lorenzo&lists=public-webagents)
- That led to this paper: Interoperability for Web Agents --> [10.36227/techrxiv.175606898.81408945/v1](https://doi.org/10.36227/techrxiv.175606898.81408945/v1), while researching for this I found the potential for a kind of vector search per statement of needs
- implementation of the paper if relevant: https://github.com/Mec-iS/bmpp-agents-rs
- Implementation for [Meaning Typed Prompting](https://github.com/Mec-iS/vibelang-rs), again this is all work connected to define RAG pipelines that involve emebddings.

From my personal perspective (besides the experience with implementing a numerical library that provided me with some foundations) this is the result of work started in June that can be tracked in public repositories and mailing lists, let me know if it qualifies. I understand the repository itself has not a long history but the research effort does.

Summarising, it just happened to come together now that I have more time to do research. I don't really know if you can consider these efforts organic to define the object of the repository itself, coming from an industry background I reason in terms of prototyping; this is an experimental prototype (POC) with testing and it is organic with what I am doing as a professional in a quite dynamic scenario. I think this research/code is very relevant as follows important current research trends (for example see [this paper](https://arxiv.org/pdf/2508.21038)); also, if you wonder about continuity, I plan to keep adding features to it to build a caching engine leveraging `smartcore`.

Anyway I don't want to overstretch so let me know if this qualifies according to your policies.
