### 2025-09-02 40e6261
add references block
