### 2025-09-02 #8893 @Mec-iS
added the affiliation details and the references block
