### 2025-09-05 dccac6b
add usage example
