### 2025-09-05 #8893 @danielskatz
@Mec-iS - the title of the paper seems odd to me, since the paper is meant to be about the software, not a primer to a field of study. Perhaps you can change it to something that better describes the software?

Also in the paper, if you refer to a reference directly, please do it without the []s around the reference. For example
* See @ref1
* using magic [@ref2]

See the example JOSS paper if needed.
