### 2025-09-10 #8893 @Mec-iS
I have changed the title to: "ArrowSpace: An Example of Spectral Indexing for vector search" to make it different from the long version.

As far a I can read, all the references are already in square brackets `[@...]`, separated by `;` when multiple are present. Please point out the line if I have missed it.

Potential reviewers from the list:
* @HadrienG
* @SebastianoF
* @johanneskoester
* @saaikrishnan
* @bricoletc
* @sstadick
* @apraga
* @remram44
In general anybody with experience in Rust, Machine Learning and Bioinformatics should find the content ready to start with.

FYI the extended version of this primer has been published in [this pre-print server](https://www.techrxiv.org/users/685780/articles/1329993-arrowspace-spectral-indexing-of-embeddings-using-taumode-%CE%BB%CF%84)

Thank you
