### 2025-09-10 #8893 @danielskatz
Please read my comment again. Some of your references should not be in square brackets, when you discuss them in the text, rather than when they are just references to text but not directly addressed.
