### 2025-09-11 e387699
change references: Narrative vs parenthetical
