### 2025-09-13 #8893 @sstadick
@danielskatz - It's a bit outside my area from a math standpoint, but I'm up for reviewing it.
