### 2025-09-16 #9002 @editorialbot
[REVIEW]: ArrowSpace: Spectral Indexing of Embeddings using taumode (λτ)
**Submitting author:** @Mec-iS (<a href="http://orcid.org/0000-0002-8804-2963">Lorenzo Moriondo</a>)
**Repository:** https://github.com/Mec-iS/arrowspace-rs
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.0-joss
