### 2025-09-16 Mec-iS/arrowspace-rs#1 @DiogoRibeiro7
Add Community Guidelines and Contribution Infrastructure
The repository is missing essential community guidelines and infrastructure required for open source projects, particularly for JOSS submission standards.

## Missing Components

### 1. Contributing Guidelines
- [ ] Create `CONTRIBUTING.md` with:
  - Development setup instructions
  - Code style guidelines (rustfmt, clippy rules)
  - Testing requirements
  - Pull request process
  - Commit message conventions

### 2. Issue Templates
- [ ] Create `.github/ISSUE_TEMPLATE/` directory with templates for:
  - **Bug Report** (`bug_report.md`)
  - **Feature Request** (`feature_request.md`)
  - **Documentation** (`documentation.md`)
  - **Performance Issue** (`performance.md`)

### 3. Pull Request Template
- [ ] Create `.github/pull_request_template.md` with:
  - Description of changes
  - Testing checklist
  - Breaking changes notification
  - Related issues linking

### 4. Support Guidelines
- [ ] Add support information to README.md:
  - How to get help
  - Where to ask questions
  - Response time expectations
  - Preferred communication channels

### 5. Code of Conduct
- [ ] Add `CODE_OF_CONDUCT.md` (recommend using Contributor Covenant)

## JOSS Requirement Context
This is required for JOSS submission under their community guidelines criteria:
> "Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support"

## Acceptance Criteria
- [ ] All community files are present and properly formatted
- [ ] Guidelines are clear and actionable for new contributors
- [ ] Templates include all necessary fields for effective issue reporting
- [ ] README.md updated with support and contribution links
- [ ] Documentation follows GitHub community standards

## Priority
**High** - Required for JOSS submission compliance

Addresses requirements from JOSS review: openjournals/joss-reviews#9002

**Labels:** `documentation`, `community`, `joss-requirement`, `good first issue`
