### 2025-09-16 Mec-iS/arrowspace-rs#2 @DiogoRibeiro7
Add Installation Instructions to Documentation
Addresses requirements from JOSS review: openjournals/joss-reviews#9002

The README.md currently lacks basic installation instructions, which is required for JOSS submission and makes it difficult for users to get started with ArrowSpace.

## Current State
The README includes:
- ✅ Project description
- ✅ Usage examples (`cargo run --example ...`)
- ✅ Benchmark commands (`cargo bench`)
- ❌ **Missing**: How to actually install/use the library

## Required Installation Documentation

### As a Library Dependency
Add to your `Cargo.toml`:
[dependencies]
arrowspace = "0.1.0"
```

### From Source
cd arrowspace-rs
cargo build --release
```

### Running Examples
cargo run --example hypergraph_showcase
cargo run --example proteins_lookup
```

### Running Tests
cargo test
```

### 2. Add Dependencies Section
- [ ] List system requirements (Rust 1.78+ is mentioned but not prominently)
- [ ] Clarify any optional dependencies
- [ ] Note any platform-specific requirements

### 3. Quick Start Example
- [ ] Add a simple "Hello World" style example showing basic usage
- [ ] Include imports and minimal working code

## JOSS Requirement Context
From JOSS review criteria:
> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

## Current User Experience Problem
New users currently see:
1. Examples that run with `cargo run --example`
2. But no guidance on how to use ArrowSpace as a library
3. No clear installation path for getting started

## Acceptance Criteria
- [ ] README.md includes comprehensive installation section
- [ ] Dependencies are clearly listed
- [ ] Multiple installation methods covered (dependency, source, development)
- [ ] Quick start example included
- [ ] Installation instructions tested on clean system

## Priority
**High** - Required for JOSS submission compliance

**Labels:** `documentation`, `joss-requirement`, `user-experience`
