### 2025-09-16 Mec-iS/arrowspace-rs#3 @DiogoRibeiro7
Fix Import Statement Capitalization in Paper
Addresses requirements from JOSS review: openjournals/joss-reviews#9002

The ArrowSpace paper contains incorrect import statements that prevent the code examples from compiling when users try to follow the documentation.

## Specific Issue

### Import Statement Capitalization Error
**Paper shows (INCORRECT):**
```rust
use ArrowSpace::builder::ArrowSpaceBuilder;  // Capital 'A'
use ArrowSpace::core::ArrowItem;            // Capital 'A'
```

**Actual code requires (CORRECT):**
```rust
use arrowspace::builder::ArrowSpaceBuilder; // lowercase 'a'
use arrowspace::core::ArrowItem;           // lowercase 'a'
```

## Impact on Users
When users copy-paste the paper examples:
1. ❌ Compilation fails with "crate not found" error
2. ❌ Users think the software is broken
3. ❌ JOSS reviewers cannot verify functionality

## The Code Example Itself is Correct
The actual usage code in the paper works fine:
```rust
let items = vec![
    vec![1.0, 2.0, 3.0], // Item 1
    vec![2.0, 3.0, 1.0], // Item 2
    vec![3.0, 1.0, 2.0], // Item 3
];

let (aspace, _graph) = ArrowSpaceBuilder::new()
    .with_rows(items)
    .with_lambda_graph(1e-3, 6, 2.0, None)
    .build();

let query = ArrowItem::new(vec![1.5, 2.5, 2.0], 0.0);
let results = aspace.search_lambda_aware(&query, 5, 1.0, 0.0);
```

**Only the imports need fixing.**

## Required Fixes

### Option A: Fix the Paper (Recommended)
- [ ] Update import statements to use lowercase `arrowspace`
- [ ] Add the corrected imports to the paper example

### Option B: Add Re-export for Compatibility
- [ ] Add to `lib.rs`:
```rust
// Re-export with capital A for paper compatibility
pub use crate as ArrowSpace;
```

## Complete Working Example for Paper
```rust
use arrowspace::builder::ArrowSpaceBuilder;  // Fixed: lowercase 'a'
use arrowspace::core::ArrowItem;            // Fixed: lowercase 'a'

// [rest of the example code remains the same]
```

## JOSS Requirement Context
From JOSS review criteria:
> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Examples must be **functional and tested**, not just illustrative.

## Acceptance Criteria
- [ ] All code examples in paper compile successfully
- [ ] Examples can be run without modification
- [ ] Import statements match actual crate structure
- [ ] Complete context provided for all examples
- [ ] Examples tested in CI/automated testing

## Priority
**High** - Blocks JOSS submission and creates poor user experience

---
**Labels:** `documentation`, `joss-requirement`, `bug`, `user-experience`
