### 2025-09-17 Mec-iS/arrowspace-rs#4 @DiogoRibeiro7
Fix Invalid Cargo.toml Configuration
Addresses requirements from JOSS review: openjournals/joss-reviews#9002

## Description
The `Cargo.toml` file contains an invalid section `[examples]` with a dependency declaration that should be in the proper `[dependencies]` or `[dev-dependencies]` section.

## Current Invalid Configuration
**In Cargo.toml (INCORRECT):**
```toml
[examples]
range-set-blaze = "^0.3.0"
```

## The Problem
- `[examples]` is not a valid Cargo.toml section for dependencies
- The `range-set-blaze` crate dependency is declared in the wrong place
- This causes confusion and potential build issues

## Valid Cargo.toml Sections for Dependencies
```toml
[dependencies]
[dev-dependencies]
[build-dependencies]
```

## Required Fix

### Option 1: Move to Dependencies (if used in main code)
```toml
[dependencies]
range-set-blaze = "0.3.0"
```

### Option 2: Move to Dev Dependencies (if used only in examples/tests)
```toml
[dev-dependencies]
range-set-blaze = "0.3.0"
```

### Option 3: Remove Entirely (if not actually used)
```toml
# Just delete the invalid section
```

## Investigation Needed
- [ ] Check if `range-set-blaze` is actually imported/used anywhere in the codebase
- [ ] Determine if it's needed for examples, tests, or main functionality
- [ ] Place in appropriate dependency section based on usage

## Impact
- **Installation issues**: Invalid Cargo.toml sections can cause build problems
- **JOSS compliance**: Reviewers expect proper Rust packaging standards
- **User confusion**: Non-standard configuration is misleading

## JOSS Requirement Context
From JOSS review criteria:
> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Dependencies must be properly declared in standard Cargo.toml sections.

## Acceptance Criteria
- [ ] `[examples]` section removed from Cargo.toml
- [ ] `range-set-blaze` dependency moved to appropriate section (or removed)
- [ ] `cargo build` and `cargo test` work without warnings
- [ ] Cargo.toml follows standard Rust packaging conventions

---
**Labels:** `build`, `configuration`, `joss-requirement`, `packaging`
