### 2025-09-17 Mec-iS/arrowspace-rs#5 @sstadick
Add MIT License File to Repo
Related to https://github.com/openjournals/joss-reviews/issues/9002

# Description

In the `Cargo.toml` the `license` fields SPDX License is `MIT OR Apache-2.0`, but there is no MIT license in the repo.

# Impact on Users

Downstream tooling like that used in Conda to automatically extract license info from Rust projects will fail to find the MIT license text.

# Priority

Low - The requirement is still met for `Does the repository contain a plain-text LICENSE or COPYING file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?`

---

**Labels**: `license`
