### 2025-09-17 c182239
fix: small readme wording suggestion
