### 2025-09-17 Mec-iS/arrowspace-rs#7 @sstadick
Fix tests
Related to https://github.com/openjournals/joss-reviews/issues/9002

# Description

`cargo build` results in a long list of warnings about unused imports in tests. This is because the `tests` module is included as part of the source and not behind a `#[cfg(test)]`, see https://doc.rust-lang.org/book/ch11-03-test-organization.html#the-tests-module-and-cfgtest, or tests could be moved to the integration test level per https://doc.rust-lang.org/book/ch11-03-test-organization.html#integration-tests, by moving the `tests` dir to the top level of the repo.

When running `cargo test` or `cargo build` there are many errors about unused imports in the tests because the imports are part of the build, but all the `#[test]` functions are not. Those functions are only compiled when running tests.

# JOSS Requirement Context

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

# Acceptance Criteria

- [ ] No warnings should be produced when running `cargo test` or `cargo build`
- [ ] Test structure should match standard Rust project tests

# Priority

Low - Technically there are tests, and they do run and pass.

---

Labels: `tests`
