+++ src/tests/mod.rs
@@ -1,7 +1,14 @@
+#[cfg(test)]
 mod test_arrow;
+#[cfg(test)]
 mod test_builder;
+#[cfg(test)]
 mod test_dimensional;
+#[cfg(test)]
 mod test_dimensional_ops;
+#[cfg(test)]
 mod test_graph_factory;
+#[cfg(test)]
 mod test_operators;
+#[cfg(test)]
 mod test_taumode;
