### 2025-09-17 Mec-iS/arrowspace-rs#7 @Mec-iS
Done 👍🏼
