### 2025-09-17 Mec-iS/arrowspace-rs#3 @Mec-iS
Done 👍🏼
