### 2025-09-17 #9002 @sstadick
@danielskatz - what's the best way (if any) to suggest writing changes? Or is that outside the scope here? Specifically in the Statement of Need I'd like to request some clarification on how spectral and semantic similarity are different, since the spectral graph is a graph of similarity (very open to being corrected on this!).

I don't want to bog things down questioning things that don't need to be questioned!
