### 2025-09-17 Mec-iS/arrowspace-rs#8 @sstadick
Formatting of Line 57
Related to https://github.com/openjournals/joss-reviews/issues/9002

All lines refer to https://raw.githubusercontent.com/openjournals/joss-papers/joss.09002/joss.09002/10.21105.joss.09002.pdf

# Description

Line 57 looks like it should be a list. There should be a newline after the `:` here: https://github.com/Mec-iS/arrowspace-rs/blob/0faf193b8ab489edd4e7ed8b48cc67c843ad74ac/paper.md?plain=1#L48

# JOSS Requirement Context

> Quality of writing: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

# Priority

High - Required for JOSS submission compliance

Labels: `formatting`
