### 2025-09-17 dd88cc0
Add contributing and project automation
