### 2025-09-17 Mec-iS/arrowspace-rs#10 @sstadick
Question: How are benchmark numbers derived?
Related to https://github.com/openjournals/joss-reviews/issues/9002
Line numbers refer to: https://raw.githubusercontent.com/openjournals/joss-papers/joss.09002/joss.09002/10.21105.joss.09002.pdf

# Description

On line 95-96 there are benchmark numbers, but it's not immediately obvious how they are derived from running `cargo bench`, or another means, in order to reproduce them.

The memory footprint makes sense, although it would be nice to see some comparison against another package (not-needed though)

# JOSS Requirement Context

>  Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

# Acceptance Criteria

- [ ] A note in the REAMDE or a some other documentation in repo about how those numbers are derived / checked.

# Suggestions (non-blocking)

- Add a plot that demonstrates the linear scaling with batch size.
- Add a plot demonstrating the 15% overhead

# Priority

High - Required for JOSS submission compliance

Labels: `documentation`, `reproducibility`
