### 2025-09-17 #9002 @danielskatz
@sstadick - thanks for pointing this out.

@sstadick - please do keep this in mind. If you have suggestions for new content, perhaps they could be in the repo and then just pointed to in the paper, or alternatively, if you suggest new material than needs to be in the paper, please think about what you would suggest removing to make space.
