### 2025-09-17 Mec-iS/arrowspace-rs#11 @sstadick
Add source of data found in `proteins_lookup.rs` example.
Related to https://github.com/openjournals/joss-reviews/issues/9002
Line numbers refer to: https://raw.githubusercontent.com/openjournals/joss-papers/joss.09002/joss.09002/10.21105.joss.09002.pdf

# Description

Line 102 of the Results section directs the reader to look at `proteins_lookup` example. I would suggest that a summary of what is there be included in the manuscript itself.

Looking at `proteins_lookup.rs`:

- Where did these "...small, normalized-ish floats to mimic compact “feature” descriptors derived from protein signals (e.g., temporal+spatial MD descriptors)." come from? What was used to create the features?

# JOSS Requirement Context

>  Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

# Acceptance Criteria

- [ ] Provide a source for the data in the examples that is referred to by the paper, ideally showing how the vectors were generated.

# Priority

High - Required for JOSS submission compliance

Labels: `documentation`, `reproducibility`
