### 2025-09-17 Mec-iS/arrowspace-rs#10 @Mec-iS
Thanks, that part has to be definitely improved.
It is limited to the scope of the current implementation that is meant to be an "example", those numbers vary a lot depending on these factors: value passed to the `build_graph_lambda` builder method and the tau and taumode extra parameters. It really depends what is the objective of the search, so if the search needs to be narrow then those parameters should be set to keep closer to the baseline behaviour; otherwise for a more "creative" search the same parameters may be set to different values. This involve different trade-offs in terms of performance and search results.
The base result from the benchmark is that querying is only 8-12% slower compared to cosine similarity while providing these adjustable parameters for enabling a broader "discovery" of the dataset.
The advantages that this approach gives are still to be investigated, that goes beyond the object of an example. It is what I am doing at the moment and I will publish some more in the next months.

That part can probably be changed from:
```
### Benchmarks

The library includes comprehensive benchmarks comparing `ArrowSpace` with baseline cosine similarity:

- **Single Query**: ~15% overhead for $λτ$-aware search vs pure cosine
- **Batch Queries**: Scales linearly with batch size, maintains constant per-query overhead
- **Memory Footprint**: 4-8 bytes per $λτ$ index vs graph storage
```
to:
```
### Benchmarks

The library includes benchmarks comparing `ArrowSpace` with baseline cosine similarity, the main benchmark shows 8-12% overhead for $λτ$-aware search compared to pure cosine. This is a low cost to pay for allowing the extension in search capabilities that the additional indexing layer enables.
```
The remaining points are more speculative at the moment.
