### 2025-09-17 Mec-iS/arrowspace-rs#10 @sstadick
I think that changing the wording like that would be totally fine and then you could just rely on `cargo bench` outputs alone.

However, when I run `cargo bench` the results look more like 100% overhead (2x slower)? (except for the single category, which is much closer but does vary a lot run to run).

```
     Running benches/lookup_bench.rs (target/release/deps/lookup_bench-adc67405ffece868)
Gnuplot not found, using plotters backend
lookup_topk_k=3/baseline_cosine/single
                        time:   [3.8661 µs 4.1126 µs 4.3794 µs]
                        change: [+7.5199% +48.907% +105.18%] (p = 0.03 < 0.05)
                        Performance has regressed.
Found 7 outliers among 40 measurements (17.50%)
  3 (7.50%) high mild
  4 (10.00%) high severe
lookup_topk_k=3/arrow_alpha1_beta0/single
                        time:   [5.5753 µs 5.9923 µs 6.5475 µs]
                        change: [+0.7293% +28.613% +72.742%] (p = 0.15 > 0.05)
                        No change in performance detected.
Found 4 outliers among 40 measurements (10.00%)
  4 (10.00%) high severe
lookup_topk_k=3/arrow_alpha0.9_beta0.1/single
                        time:   [5.6781 µs 6.1200 µs 6.5952 µs]
                        change: [−10.915% −3.2777% +3.5336%] (p = 0.44 > 0.05)
                        No change in performance detected.
Found 3 outliers among 40 measurements (7.50%)
  1 (2.50%) high mild
  2 (5.00%) high severe
lookup_topk_k=3/baseline_cosine/batch16
                        time:   [31.263 µs 31.981 µs 32.849 µs]
                        change: [−0.6948% +3.6606% +8.5816%] (p = 0.11 > 0.05)
                        No change in performance detected.
Found 4 outliers among 40 measurements (10.00%)
  2 (5.00%) high mild
  2 (5.00%) high severe
lookup_topk_k=3/arrow_alpha1_beta0/batch16
                        time:   [61.041 µs 61.771 µs 62.563 µs]
                        change: [−0.8471% +0.9790% +2.7405%] (p = 0.30 > 0.05)
                        No change in performance detected.
Found 4 outliers among 40 measurements (10.00%)
  2 (5.00%) high mild
  2 (5.00%) high severe
lookup_topk_k=3/arrow_alpha0.9_beta0.1/batch16
                        time:   [62.258 µs 63.556 µs 65.070 µs]
                        change: [−0.5018% +1.3961% +3.7626%] (p = 0.19 > 0.05)
                        No change in performance detected.
Found 3 outliers among 40 measurements (7.50%)
  3 (7.50%) high severe
lookup_topk_k=3/baseline_cosine/batch64
                        time:   [111.36 µs 112.52 µs 113.90 µs]
                        change: [−2.5385% −0.5450% +1.3631%] (p = 0.60 > 0.05)
                        No change in performance detected.
Found 3 outliers among 40 measurements (7.50%)
  2 (5.00%) high mild
  1 (2.50%) high severe
lookup_topk_k=3/arrow_alpha1_beta0/batch64
                        time:   [233.81 µs 235.89 µs 237.88 µs]
                        change: [−0.2880% +0.6231% +1.5436%] (p = 0.18 > 0.05)
                        No change in performance detected.
Found 3 outliers among 40 measurements (7.50%)
  2 (5.00%) high mild
  1 (2.50%) high severe
lookup_topk_k=3/arrow_alpha0.9_beta0.1/batch64
                        time:   [241.77 µs 243.87 µs 246.42 µs]
                        change: [−1.2268% +0.3474% +1.7577%] (p = 0.67 > 0.05)
                        No change in performance detected.
Found 7 outliers among 40 measurements (17.50%)
  3 (7.50%) high mild
  4 (10.00%) high severe
lookup_topk_k=3/baseline_cosine/batch128
                        time:   [223.26 µs 225.89 µs 228.86 µs]
                        change: [−1.6839% +0.1381% +1.8348%] (p = 0.89 > 0.05)
                        No change in performance detected.
Found 1 outliers among 40 measurements (2.50%)
  1 (2.50%) high severe
lookup_topk_k=3/arrow_alpha1_beta0/batch128
                        time:   [466.57 µs 469.25 µs 472.75 µs]
                        change: [+0.0773% +1.1570% +2.1059%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 40 measurements (7.50%)
  3 (7.50%) high severe
lookup_topk_k=3/arrow_alpha0.9_beta0.1/batch128
                        time:   [482.72 µs 489.81 µs 499.94 µs]
                        change: [+0.3165% +1.4895% +2.8066%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 40 measurements (7.50%)
  1 (2.50%) high mild
  2 (5.00%) high severe
lookup_topk_k=3/baseline_cosine/batch256
                        time:   [447.87 µs 453.52 µs 460.56 µs]
                        change: [−13.606% −3.9571% +2.3625%] (p = 0.64 > 0.05)
                        No change in performance detected.
Found 2 outliers among 40 measurements (5.00%)
  1 (2.50%) high mild
  1 (2.50%) high severe
lookup_topk_k=3/arrow_alpha1_beta0/batch256
                        time:   [925.43 µs 932.14 µs 941.77 µs]
                        change: [−8.4169% −4.5983% −1.3891%] (p = 0.01 < 0.05)
                        Performance has improved.
Found 4 outliers among 40 measurements (10.00%)
  2 (5.00%) high mild
  2 (5.00%) high severe
lookup_topk_k=3/arrow_alpha0.9_beta0.1/batch256
                        time:   [959.31 µs 1.0179 ms 1.1140 ms]
                        change: [+0.8448% +3.9088% +10.206%] (p = 0.06 > 0.05)
                        No change in performance detected.
Found 4 outliers among 40 measurements (10.00%)
  2 (5.00%) high mild
  2 (5.00%) high severe
```

on a MacBook Pro M1. Entirely possible I'm misreading things though!
