### 2025-09-17 Mec-iS/arrowspace-rs#10 @Mec-iS
I just realise now that these benchmarks are not very explicative as they use `lambda_similarity` instead of the actual method to benchmark for a vectordb-like application that should be `search_lambda_aware`. So basically they are testing the index-building capabilities and not the search capability. I guess I forgot to refactor the bench after the latest API change.

I need to refactor the bench, this is the `index_compute_bench`, not the `lookup_bench`. So I guess I have to build the lookup bench. Thanks for asking this.
