### 2025-09-17 Mec-iS/arrowspace-rs#12 @DiogoRibeiro7
Example `hypergraph_showcase` Not Found - Documentation Update Needed
Addresses requirements from JOSS review: openjournals/joss-reviews#9002

## Description
When attempting to run the `hypergraph_showcase` example as documented, the command fails with an error indicating that the example target does not exist.

## Steps to Reproduce
1. Navigate to the project directory
2. Run the command: `cargo run --example hypergraph_showcase`

## Expected Behavior
The example should execute successfully, demonstrating hypergraph functionality.

## Actual Behavior
```
warning: C:\Users\diogo\work_code\arrowspace-rs\Cargo.toml: unused manifest key: examples
error: no example target named `hypergraph_showcase` in default-run packages
help: available example targets:
    compare_cosine
    merge_signals
    proteins_lookup
```

## Environment
- OS: Windows
- Project path: `C:\Users\diogo\work_code\arrowspace-rs\`

## Additional Notes
The error message indicates that the following examples are currently available:
- `compare_cosine`
- `merge_signals`
- `proteins_lookup`

There's also a warning about an unused manifest key `examples` in the Cargo.toml file.

## Suggested Resolution
Either:
1. Add the missing `hypergraph_showcase` example implementation, or
2. Update the documentation to reference one of the existing examples instead
3. Fix the unused manifest key warning in Cargo.toml
