### 2025-09-17 Mec-iS/arrowspace-rs#13 @DiogoRibeiro7
Fix Community Guidelines File Structure and Organization
Addresses requirements from JOSS review: openjournals/joss-reviews#9002

## Problem
The repository's community guidelines are incorrectly placed and incompletely structured, which prevents GitHub from properly detecting and surfacing them to potential contributors.

## Current Issues

### 1. Incorrect File Placement ❌
- `CONTRIBUTING.md` is in `.github/` instead of repository root
- `CODE_OF_CONDUCT.md` is in `.github/` instead of repository root
- Single generic `ISSUE_TEMPLATE.md` instead of organized template directory

### 2. Missing Specialized Templates ❌
- No categorized issue templates for different types of contributions
- Generic template doesn't guide users to provide appropriate information for different issue types

## Expected Structure ✅

The community guidelines should follow GitHub's conventions:

```
├── CONTRIBUTING.md              ← Move from .github/
├── CODE_OF_CONDUCT.md           ← Move from .github/
└── .github/
    ├── ISSUE_TEMPLATE/
    │   ├── bug_report.yml       ← New: Structured bug reports
    │   ├── feature_request.yml  ← New: Feature requests
    │   ├── documentation.yml    ← New: Documentation improvements
    │   └── performance.yml      ← New: Performance issues
    ├── pull_request_template.md
    └── CODEOWNERS
```

## Benefits of Fix

1. **GitHub Integration**: Files will be automatically detected and linked in GitHub UI
2. **Better Contributor Experience**: Clear, categorized templates guide better issue reports
3. **Maintainer Efficiency**: Structured templates provide consistent information for triage
4. **Community Standards**: Follows established GitHub repository conventions

## Proposed Changes

### Phase 1: Move Core Files
- [ ] Move `CONTRIBUTING.md` to repository root
- [ ] Move `CODE_OF_CONDUCT.md` to repository root
- [ ] Update any internal links that reference the old locations

### Phase 2: Create Structured Issue Templates
- [ ] Create `.github/ISSUE_TEMPLATE/` directory
- [ ] Convert current `ISSUE_TEMPLATE.md` to structured YAML templates:
  - `bug_report.yml` - For bug reports with system info, steps to reproduce
  - `feature_request.yml` - For new feature suggestions with use cases
  - `documentation.yml` - For documentation improvements
  - `performance.yml` - For performance-related issues with benchmarks
- [ ] Remove old generic template

### Phase 3: Enhance Templates
- [ ] Add relevant fields for each template type (environment details, expected behavior, etc.)
- [ ] Include links to contribution guidelines in templates
- [ ] Add appropriate labels and assignees for automatic triage

- [[GitHub Documentation: Setting guidelines for repository contributors](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)
- [[GitHub Documentation: Configuring issue templates](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository)](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository)
