### 2025-09-17 Mec-iS/arrowspace-rs#15 @DiogoRibeiro7
Fix Performance Claims - Benchmarks Show 35-160% Overhead, Not 15%
Addresses requirements from JOSS review: https://github.com/openjournals/joss-reviews/issues/9002

## Problem
Documentation claims "15% overhead" but `cargo bench` shows 35-160% overhead vs baseline cosine similarity.

## Benchmark Results

| Scenario       | Baseline | ArrowSpace | Actual Overhead |
|----------------|----------|------------|-----------------|
| Single query   | 4.15 µs  | 5.996 µs   | **44%**         |
| 256 queries    | 471 µs   | 1170 µs    | **148%**        |

## Required Fixes

### 1. Update Documentation

- [ ] Fix "15% overhead" claim in README.md
- [ ] Add honest performance section: "35-45% overhead for single queries, 140-160% for batch queries"
- [ ] Explain trade-off: "Trades speed for spectral similarity insights"

### 2. Add Competitive Benchmarks

- [ ] Compare against FAISS, Annoy, scikit-learn
- [ ] Test larger datasets (currently only 64 items)
- [ ] Add memory usage measurements

### 3. Expand Test Coverage

- [ ] Test 10K, 100K+ vectors
- [ ] Multiple dimensions (128, 512, 1024)
- [ ] CI performance regression tests

## Expected Outcome

Users get accurate performance expectations and can make informed decisions about when ArrowSpace's spectral features are worth the performance trade-off.
