### 2025-09-17 Mec-iS/arrowspace-rs#16 @DiogoRibeiro7
Improve JOSS Paper Writing Quality and Accessibility
Addresses requirements from JOSS review: https://github.com/openjournals/joss-reviews/issues/9002

## Problem
The `paper.md` follows correct JOSS structure but has writing quality issues that may affect review. The paper is too technical for JOSS's broad software developer audience and contains grammar/clarity issues.

## Current Issues

### 1. **Summary Too Technical** ❌
```markdown
# Current (hard to understand):
"`ArrowSpace` provides an API to use taumode (`λτ`) that is a single, bounded synthetic
score per signal that blends the Rayleigh smoothness energy on a graph with an edgewise
dispersion summary"

# Should be (accessible):
"ArrowSpace is a Rust library for vector similarity search that goes beyond traditional
distance metrics by incorporating spectral graph properties to find structural patterns
in high-dimensional data."
```

### 2. **Inconsistent Terminology** ❌
- Mixes "`λτ`" and "lambda-tau" throughout
- "E'" vs "E′" formatting inconsistency
- "taumode" vs "λτ indexing" terminology

### 3. **Grammar and Style Issues** ❌
```markdown
# Fix grammar:
"From an engineering perspective, there is increasing demand..."
→ "Vector databases increasingly need methods beyond geometric similarity..."

"For this research LLMs have been used extensively in the ideation and development phase."
→ Move to acknowledgments or remove
```

## Suggested Improvements

### **High Priority** - Required for JOSS

#### 1. Simplify Summary Section
```markdown
# Summary

ArrowSpace is a Rust library for vector similarity search that incorporates spectral
graph properties alongside traditional geometric measures. While traditional vector
databases rely on distance metrics like cosine similarity, ArrowSpace adds a spectral
dimension that captures structural patterns, enabling more nuanced similarity matching
for scientific and structured data applications.

The library introduces λτ (lambda-tau) indexing, which combines Rayleigh quotient
energy with dispersion statistics to create bounded similarity scores. This approach
is particularly valuable for domains where geometric similarity alone is insufficient,
such as protein analysis, signal processing, and molecular dynamics.
```

#### 2. Clarify Statement of Need
```markdown
# Statement of Need

Traditional vector similarity search relies on geometric measures like cosine similarity
or Euclidean distance. These methods capture semantic relationships but ignore spectral
structure—the patterns in how data points relate to each other across the entire dataset.

For scientific applications like protein analysis or signal processing, this limitation
means that structurally similar samples may be ranked poorly by traditional distance
metrics, while geometrically close but structurally different samples rank highly.

ArrowSpace addresses this gap by providing the first integrated spectral-aware indexing
for vector databases, enabling similarity search that considers both content and
structural context.
```

#### 3. Standardize Terminology
- Use **"λτ indexing"** consistently throughout
- Standardize mathematical notation (choose E' or E′)
- Define technical terms briefly when first used

###  Polish improvements

#### 4. Grammar and Style Fixes
- [ ] Fix awkward sentence constructions
- [ ] Improve flow between paragraphs
- [ ] Remove/relocate LLM disclaimer
- [ ] Consistent bullet point formatting
