### 2025-09-17 Mec-iS/arrowspace-rs#17 @DiogoRibeiro7
Minor JOSS Paper Improvements - Strengthen State of the Field Context
Addresses requirements from JOSS review: https://github.com/openjournals/joss-reviews/issues/9002

## Background
The current `paper.md` meets JOSS requirements for state of the field, but could benefit from minor enhancements to help reviewers and users better understand ArrowSpace's positioning relative to existing tools.

## Current Status: ✅ ACCEPTABLE
JOSS doesn't require extensive literature reviews or comprehensive benchmarking like academic journals. The current paper adequately:
- Identifies the problem (spectral structure ignored by traditional methods)
- Explains the gap (no integrated spectral-aware indexing)
- Provides basic theoretical context

## Suggested Minor Improvements

### 1. Add Brief Comparison Section (Optional)
Add 2-3 sentences comparing with major alternatives:

```markdown
## Relationship to Existing Software
ArrowSpace differs from existing vector search libraries such as FAISS [@Johnson:2017],
Annoy, and scikit-learn by incorporating spectral graph properties alongside traditional
geometric similarity. While these libraries optimize for query speed, ArrowSpace trades
performance (35-60% overhead) for spectral similarity insights.
```

### 2. Clarify Performance Trade-offs (Helpful)
Update the existing performance section to be more explicit:

```markdown
## Performance Characteristics
ArrowSpace trades query speed for spectral insights, showing 35-60% overhead versus
pure cosine similarity for single queries. This trade-off is justified in domains
where structural patterns matter more than raw query speed.
```

### 3. Strengthen Use Case Positioning (User Guidance)
Make the target applications clearer:

```markdown
## Target Applications
The library is specifically designed for scientific computing domains where geometric
similarity alone is insufficient, including protein dynamics analysis, signal processing,
and molecular structure comparison.
```

## Benefits of These Changes
- **Reviewer clarity**: Helps JOSS reviewers understand positioning vs alternatives
- **User guidance**: Makes it clearer when ArrowSpace is the right choice
- **Honest expectations**: Sets realistic performance expectations

## Implementation
- [ ] Add 1-2 paragraph "Relationship to Existing Software" section
- [ ] Update performance claims to be more explicit about trade-offs
- [ ] Clarify target scientific domains in summary/introduction
