### 2025-09-18 Mec-iS/arrowspace-rs#18 @DiogoRibeiro7
Complete JOSS Paper References and Citations
Addresses requirements from JOSS review: https://github.com/openjournals/joss-reviews/issues/9002

## Problem
The `paper.md` has correct citation syntax but is missing several important references that JOSS reviewers expect, particularly software dependencies and key related work mentioned in the text.

## Missing Citations

### 1. **Software Dependencies**
The paper uses but doesn't cite key dependencies from `Cargo.toml`:

**Add to paper.bib:**
```bibtex
@software{smartcore2021,
  author = {Orlov, Vlad},
  title = {SmartCore: Machine Learning Library for Rust},
  url = {https://smartcorelib.org/},
  year = {2019}
}

@software{rust2015,
  author = {{Rust Team}},
  title = {The Rust Programming Language},
  url = {https://www.rust-lang.org/},
  year = {2015},
  note = {Edition 2024}
}
```

**Add to paper text:**
```markdown
ArrowSpace is implemented in Rust [@rust2015] and builds on the SmartCore [@smartcore2021]
machine learning ecosystem.
```

### 2. **Improve Repository Citation Format**
The ArrowSpace repository citation could follow standard software citation format:

```bibtex
# Current (acceptable):
@misc{ArrowSpace:2025,
    Author = {{Moriondo}, Lorenzo},
    Title = {{ArrowSpace-rs: Spectral Vector Search with Lambda-Tau Indexing}},
    Year = 2025,
    Note = {Rust codebase for spectral indexing},
    Url = {https://github.com/Mec-iS/arrowspace-rs}
}

# Optional improvement (more standard):
@software{ArrowSpace:2025,
    Author = {Moriondo, Lorenzo},
    Title = {{ArrowSpace}: Spectral Vector Search with Lambda-Tau Indexing},
    Year = {2025},
    Url = {https://github.com/Mec-iS/arrowspace-rs},
    Version = {0.1.0}
}
```

### 3. **Fix Incomplete References**
Several references lack proper publication details:

```bibtex
# Current (incomplete):
@misc{Mahadevan:2006,
    Title = {{CMPSCI 791BB: Advanced ML - Spectral Graph Theory}},
    Institution = {University of Massachusetts, Amherst},
    Year = 2006,
}

# Should be (complete):
@misc{Mahadevan:2006,
    Author = {Mahadevan, Sridhar},
    Title = {{Spectral Graph Theory Lecture Notes}},
    Institution = {University of Massachusetts, Amherst},
    Year = {2006},
    Note = {Course CMPSCI 791BB},
    Url = {https://people.cs.umass.edu/~mahadeva/cs791bb/lectures-s2006/lec4.pdf}
}
```

### 4. **Optional: Key Related Work**
For stronger positioning, consider adding major vector search references:

```bibtex
@article{johnson2017billion,
  title={Billion-scale similarity search with {GPUs}},
  author={Johnson, Jeff and Douze, Matthijs and J{\'e}gou, Herv{\'e}},
  journal={IEEE Transactions on Big Data},
  volume={7},
  number={3},
  pages={535--547},
  year={2021},
  doi={10.1109/TBDATA.2019.2921572}
}

@article{malkov2018efficient,
  title={Efficient and robust approximate nearest neighbor search using Hierarchical Navigable Small World graphs},
  author={Malkov, Yu A and Yashunin, D A},
  journal={IEEE transactions on pattern analysis and machine intelligence},
  volume={42},
  number={4},
  pages={824--836},
  year={2020},
  doi={10.1109/TPAMI.2018.2889473}
}
```
