### 2025-09-18 3add79f
add corrections to benchmarks section and acknoledgements
