+++ src/tests/test_random_walk_molecules.rs
@@ -1,245 +1,0 @@
-//! Test module: free random walk molecular diffusion simulation
-//! Reference: Physical Models of Living Systems, Chapter 6
-//! Simulates Brownian motion of a molecule and computes Arrow lambda on the trajectory
-//!
-//! Arrow Space Integration
-//! * Uses your existing build_knn_graph and ArrowSpace implementations
-//! * Computes spectral lambda on different physical signals over trajectory graphs
-//! * Tests superposition behavior (molecular interaction effects)
-//!
-//! * Validates lambda positivity and bounds for physical consistency
-//! Biological/Physical Insights
-//! * Lambda as smoothness measure: Lower lambda indicates more homogeneous concentration
-//! * Superposition effects: Combined molecular fields can have non-additive spectral properties
-//! * Time evolution: Tracks how diffusion spreads and its spectral signature changes
-
-#[cfg(test)]
-mod free_random_walk_tests {
-    use crate::core::ArrowSpace;
-    use crate::operators::{build_knn_graph};
-    use rand::prelude::*;
-
-    /// Simulate 2D free random walk of one particle on integer lattice
-    /// Returns sequence of visited coordinates
-    fn simulate_free_random_walk(steps: usize, start: (usize, usize)) -> Vec<(usize, usize)> {
-        let mut rng = rand::rng();
-        let mut pos = start;
-        let mut trajectory = Vec::with_capacity(steps);
-        trajectory.push(pos);
-
-        for _ in 1..steps {
-            // Random step in 4 directions: up, down, left, right
-            let direction = rng.random_range(0..4);
-            pos = match direction {
-                0 => (pos.0.wrapping_add(1), pos.1), // right
-                1 => (pos.0.wrapping_sub(1), pos.1), // left
-                2 => (pos.0, pos.1.wrapping_add(1)), // up
-                _ => (pos.0, pos.1.wrapping_sub(1)), // down
-            };
-            trajectory.push(pos);
-        }
-        trajectory
-    }
-
-    /// Generate concentration field from random walk trajectory
-    /// Models local concentration as function of visit frequency
-    fn concentration_field(trajectory: &[(usize, usize)]) -> Vec<f64> {
-        use std::collections::HashMap;
-
-        // Count visits to each position
-        let mut visit_counts: HashMap<(usize, usize), usize> = HashMap::new();
-        for &pos in trajectory {
-            *visit_counts.entry(pos).or_insert(0) += 1;
-        }
-
-        // Create concentration signal: log(1 + visit_count)
-        let mut signal = Vec::with_capacity(trajectory.len());
-        for &pos in trajectory {
-            let count = visit_counts[&pos] as f64;
-            signal.push((1.0 + count).ln());
-        }
-        signal
-    }
-
-    /// Generate diffusion potential field (distance from starting point)
-    fn diffusion_potential(trajectory: &[(usize, usize)], start: (usize, usize)) -> Vec<f64> {
-        trajectory
-            .iter()
-            .map(|&(x, y)| {
-                let dx = (x as f64) - (start.0 as f64);
-                let dy = (y as f64) - (start.1 as f64);
-                (dx * dx + dy * dy).sqrt()
-            })
-            .collect()
-    }
-
-    #[test]
-    fn test_molecular_diffusion_single_particle() {
-        let steps = 1000;
-        let start_pos = (50, 50);
-
-        // Simulate random walk
-        let trajectory = simulate_free_random_walk(steps, start_pos);
-
-        // Build spatial graph from trajectory points
-        let graph = build_knn_graph(&trajectory, 6, 2.0, None);
-
[…]
