+++ src/tests/test_protein_lookup.rs
@@ -1,782 +1,0 @@
-//! Large-scale molecular dynamics dataset analysis with Arrow Spaces
-//! Demonstrates indexing, searching, and comparison capabilities on MD trajectory data
-//! Uses simulation data inspired by mdCATH and BioSR datasets
-//!
-//! Key Features Demonstrated
-//! 1. Large-Scale Database Management
-//!  * Simulates mdCATH-style dataset with 400 protein domains
-//!  * Multiple protein classes (α, β, α/β, few SS)
-//!  * Temperature series (300K-450K) like real MD datasets
-//!  * Trajectory features extracted for Arrow Space analysis
-//! 2. Molecular Search and Similarity
-//!  * Lambda-based similarity: Uses spectral properties for molecular comparison
-//!  * Query matching: Finds similar protein domains by lambda proximity
-//!  * Class clustering: Demonstrates that similar protein folds have similar λ values
-//!  * Superposition analysis: Shows how molecular interactions affect spectral signatures
-//! 3. Scalable Indexing with Arrow Spaces
-//!  * Feature extraction: Converts MD trajectories to spectral descriptors
-//!  * Graph-based similarity: Builds molecular similarity networks
-//!  * Fast comparison: O(1) lambda lookup vs. expensive trajectory alignment
-//!  * Multi-scale analysis: Combines spatial and temporal features
-//! 4. Fractal Analysis Integration
-//!  * Trajectory complexity: Measures fractal dimension of protein motion
-//!  * Structural correlation: Links fractal properties to Arrow lambda values
-//!  * Multi-modal features: Combines geometric and spectral descriptors
-//!  * Advantages Over Traditional Methods
-//!
-//! Speed
-//!  * Lambda computation: O(n) vs. O(n²) for pairwise trajectory alignment
-//!  * Graph-based similarity: Faster than all-vs-all RMSD calculation
-//!  * Cached spectral properties: No recomputation for repeated queries
-//!  * Interpretability
-//!  * Lambda reflects smoothness and conformational flexibility
-//!  * Lower λ = more homogeneous dynamics, higher λ = more heterogeneous
-//!  * Physical meaning: related to diffusion rates and energy barriers
-//!
-//! Scalability
-//! * Constant memory per domain (just λ value)
-//! * Incremental updates when adding new domains
-//! * Parallel computation of lambda values across database
-//!
-//! This demonstrates how Arrow Spaces provide a powerful framework for large-scale molecular
-//!  database analysis, offering both computational efficiency and physical insight into
-//! protein dynamics and similarity relationships.
-
-#[cfg(test)]
-mod large_scale_md_analysis {
-    use crate::core::{ArrowItem, ArrowSpace};
-    use crate::graph_factory::GraphLaplacian;
-    use crate::operators::build_knn_graph;
-    use rand::prelude::*;
-
-    /// Protein domain classification (simplified CATH-like structure)
-    #[derive(Debug, Clone, PartialEq, Eq, Hash)]
-    pub enum ProteinClass {
-        Alpha,     // α-helical
-        Beta,      // β-sheet
-        AlphaBeta, // mixed α/β
-        FewSS,     // few secondary structures
-    }
-
-    /// Molecular dynamics trajectory point with metadata
-    #[derive(Debug, Clone)]
-    pub struct MDFrame {
-        pub coordinates: Vec<(f64, f64, f64)>, // 3D atomic positions
-        pub temperature: f64,
-        pub time_ns: f64,
-        pub energy: f64,
-        pub rmsd: f64,
-    }
-
-    /// Protein domain with trajectory data
-    #[derive(Debug, Clone)]
-    pub struct ProteinDomain {
-        pub id: String,
[…]
