+++ .github/ISSUE_TEMPLATE.md
@@ -1,25 +1,0 @@
-### I'm submitting a 
-- [ ] bug report.
-- [ ] improvement.
-- [ ] feature request.
-
-### Current Behaviour:
-<!-- Describe about the bug -->
-
-### Expected Behaviour:
-<!-- Describe what will happen if bug is removed -->
-
-### Steps to reproduce:
-<!-- If you can then please provide the steps to reproduce the bug -->
-
-### Snapshot:
-<!-- If you can then please provide the screenshot of the issue you are facing -->
-
-### Environment:
-<!-- Please provide the following environment details if relevant --> 
-* rustc version
-* cargo version
-* OS details
-
-### Do you want to work on this issue?
-<!-- yes/no -->

[…]
