### 2025-09-20 b8d43ec
Improve README and motivations section
