+++ paper.bib
@@ -43,6 +43,16 @@ z
     Note = {Available at: \url{https://www.cs.yale.edu/homes/spielman/462/2007/lect7-07.pdf}}
 }
 
+@article{Weller:2025,
+  title={On the Theoretical Limitations of Embedding-Based Retrieval},
+  author={Orion Weller and Michael Boratko and Iftekhar Naim and Jinhyuk Lee},
+  year={2025},
+  eprint={2508.21038},
+  archivePrefix={arXiv},
+  primaryClass={cs.IR},
+  url={https://arxiv.org/abs/2508.21038}
+}
+
 @article{Kelly:2020,
     Author = {{Kelly}, Terence},
     Title = {{Compressed Sparse Row Format for Representing Graphs}},
