### 2025-09-20 a7be03c
Add results of compare_cosine. Add reference to data sources
