+++ examples/proteins_lookup.rs
@@ -20,6 +20,11 @@ use ordered_float::OrderedFloat;
 use arrowspace::builder::{ArrowSpaceBuilder, transpose_nf_to_fn};
 use std::collections::BTreeMap;
 
+
+//
+// The data has been generated by an LLM taking as input proteins embeddings generated from various datasets in
+// https://alphafold.ebi.ac.uk/search/sequence/MENFQKVEKIGEGTYGVVYKARNKLTGEVVALKKIRLDTETEGVPSTAIREIS
+//
 // ------------ Data block: 64 items x 24 dims ------------
 // Each line: id; f0,f1,...,f23
 // These are small, normalized-ish floats to mimic compact “feature” descriptors
