### 2025-09-20 Mec-iS/arrowspace-rs#10 @Mec-iS
please see changes at "Benchmarks" section. Please reopen this issue if not satisfactory.
