### 2025-09-20 Mec-iS/arrowspace-rs#17 @Mec-iS
I decided to not add these paragraphs and some of the content is already addressed in other paragraphs and also the number of words is now ~1500 with the latest additions.
