+++ paper.bib
@@ -1,4 +1,4 @@
-@misc{ArrowSpace:2025,
+@software{ArrowSpace:2025,
     Author = {{Moriondo}, Lorenzo},
     Title = {{ArrowSpace-rs: Spectral Vector Search with Lambda-Tau Indexing}},
     Year = 2025,
@@ -6,12 +6,24 @@
     Url = {https://github.com/Mec-iS/arrowspace-rs}
 }
 
+@article{malkov:2018,
+  title={Efficient and robust approximate nearest neighbor search using Hierarchical Navigable Small World graphs},
+  author={Malkov, Yu A and Yashunin, D A},
+  journal={IEEE transactions on pattern analysis and machine intelligence},
+  volume={42},
+  number={4},
+  pages={824--836},
+  year={2020},
+  doi={10.1109/TPAMI.2018.2889473}
+}
+
 @misc{Mahadevan:2006,
-    Author = {{Mahadevan}, Sridhar},
-    Title = {{CMPSCI 791BB: Advanced ML - Spectral Graph Theory}},
+    Author = {Mahadevan, Sridhar},
+    Title = {{Spectral Graph Theory Lecture Notes}},
     Institution = {University of Massachusetts, Amherst},
-    Year = 2006,
-    Note = {Available at: \url{https://people.cs.umass.edu/~mahadeva/cs791bb/lectures-s2006/lec4.pdf}}
+    Year = {2006},
+    Note = {Course CMPSCI 791BB},
+    Url = {https://people.cs.umass.edu/~mahadeva/cs791bb/lectures-s2006/lec4.pdf}
 }
 
 @misc{Chen:2020,
@@ -90,11 +102,19 @@ z
     Year = 2010
 }
 
-@misc{smartcore:2021,
-    Author = {{Orlov}, Vlad},
-    Title = {{Smartcore: A numerical library in pure Rust}},
-    Year = 2019,
-    Note = {Available at: \url{https://github.com/smartcorelib/smartcore}}
+@software{smartcore2021,
+  author = {Orlov, Vlad},
+  title = {SmartCore: Machine Learning Library for Rust},
+  url = {https://smartcorelib.org/},
+  year = {2019}
+}
+
+@software{rust2015,
+  author = {{Rust Team}},
+  title = {The Rust Programming Language},
+  url = {https://www.rust-lang.org/},
+  year = {2015},
+  note = {Edition 2024}
 }
 
 @book{Nelson:2015,
