### 2025-09-20 10e9521
remove duplicate paragraph
