+++ paper.bib
@@ -102,7 +102,7 @@ z
     Year = 2010
 }
 
-@software{smartcore2021,
+@software{smartcore:2021,
   author = {Orlov, Vlad},
   title = {SmartCore: Machine Learning Library for Rust},
   url = {https://smartcorelib.org/},
