### 2025-09-20 #9002 @DiogoRibeiro7
@danielskatz

After reviewing the latest changes, I'm pleased to close my review and recommend this paper for publication.

The author have adequately addressed the key concerns regarding field positioning, writing quality, and reference completeness.

The paper now meets JOSS publication standards and makes a valuable contribution to the field of spectral-aware vector indexing.
