### 2025-09-25 0863824
Update paper.bib
