+++ paper.bib
@@ -124,3 +124,17 @@ z
     Year = 2015,
     Note = {Available at: \url{https://www.physics.upenn.edu/biophys/PMLS/}}
 }
+
+@article{YouKisung:2025,
+    title={Semantics at an Angle: When Cosine Similarity Works Until It Doesn't},
+    author={You, Kisung},
+    journal={arXiv preprint arXiv:2504.16318},
+    year={2025},
+    month={5},
+    note={arXiv:2504.16318v2 [cs.LG]},
+    doi={10.48550/arXiv.2504.16318},
+    url={https://arxiv.org/abs/2504.16318v2},
+    archivePrefix={arXiv},
+    eprint={2504.16318},
+    primaryClass={cs.LG}
+}
