### 2025-09-25 7b4f479
Add paper reference and adjust title
