### 2025-09-26 34ce5cc
Update paper.bib
