+++ paper.bib
@@ -4,6 +4,7 @@
     Year = 2025,
     Note = {Rust codebase for spectral indexing},
     Url = {https://github.com/Mec-iS/arrowspace-rs}
+    doi= {10.36227/techrxiv.175751921.18542359}
 }
 
 @article{malkov:2018,
