### 2025-09-26 99c4165
Update paper.bib
