### 2025-09-26 0ddac40
Update paper.bib
