+++ paper.bib
@@ -3,7 +3,7 @@
     Title = {{ArrowSpace-rs: Spectral Vector Search with Lambda-Tau Indexing}},
     Year = 2025,
     Note = {Rust codebase for spectral indexing},
-    Url = {https://github.com/Mec-iS/arrowspace-rs}
+    Url = {https://github.com/Mec-iS/arrowspace-rs},
     doi = {10.36227/techrxiv.175751921.18542359}
 }
 
