### 2025-09-26 28f2181
Update paper.bib
