+++ paper.bib
@@ -100,7 +100,8 @@ z
     Title = {{Heat Kernels, Manifolds and Graph Embedding}},
     Journal = {Pattern Recognition},
     Note = {University of California, Davis},
-    Year = 2010
+    Year = 2010,
+    doi = {10.1007/978-3-540-27868-9_20}
 }
 
 @software{smartcore:2021,
