### 2025-09-26 #9002 @danielskatz
👋 @Mec-iS - now there is an invalid DOI...  Can you try to fix this?
