### 2025-09-26 6c2aefc
Update paper.bib
