+++ paper.bib
@@ -4,7 +4,7 @@
     Year = 2025,
     Note = {Rust codebase for spectral indexing},
     Url = {https://github.com/Mec-iS/arrowspace-rs},
-    doi = {10.36227/techrxiv.175751921.18542359}
+    doi = {10.36227/techrxiv.175751921.18542359/v1}
 }
 
 @article{malkov:2018,
