### 2025-09-26 39ef91d
changes for JOSS publication (#19)
* Update paper.md
* Update paper.bib
