+++ paper.bib
@@ -1,6 +1,6 @@
 @software{ArrowSpace:2025,
     Author = {{Moriondo}, Lorenzo},
-    Title = {{ArrowSpace-rs: Spectral Vector Search with Lambda-Tau Indexing}},
+    Title = {{ArrowSpace-rs}: Spectral Vector Search with Lambda-Tau Indexing},
     Year = 2025,
     Note = {Rust codebase for spectral indexing},
     Url = {https://github.com/Mec-iS/arrowspace-rs},
@@ -20,7 +20,7 @@
 
 @misc{Mahadevan:2006,
     Author = {Mahadevan, Sridhar},
-    Title = {{Spectral Graph Theory Lecture Notes}},
+    Title = {Spectral Graph Theory Lecture Notes},
     Institution = {University of Massachusetts, Amherst},
     Year = {2006},
     Note = {Course CMPSCI 791BB},
@@ -29,7 +29,7 @@
 
 @misc{Chen:2020,
     Author = {{Chen}, Guangliang},
-    Title = {{Math 253: Rayleigh Quotient Lecture Notes}},
+    Title = {Math 253: {R}ayleigh Quotient Lecture Notes},
     Institution = {San Jose State University},
     Year = 2020,
     Note = {Available at: \url{https://www.sjsu.edu/faculty/guangliang.chen/Math253S20/lec4RayleighQuotient.pdf}}
@@ -37,7 +37,7 @@
 z
 @article{Lewis:2020,
     author = {{Lewis}, Patrick and {Perez}, Ethan and {Piktus}, Aleksandra and {Petroni}, Fabio and {Karpukhin}, Vladimir and {Goyal}, Naman and {Küttler}, Heinrich and {Lewis}, Mike and {Yih}, Wen-tau and {Rocktäschel}, Tim and {Riedel}, Sebastian and {Kiela}, Douwe},
-    title = {{Retrieval-Augmented Generation for Knowledge-Intensive NLP Tasks}},
+    title = {Retrieval-Augmented Generation for Knowledge-Intensive NLP Tasks},
     journal = {arXiv preprint},
     archivePrefix = {arXiv},
     eprint = {2005.11401},
@@ -50,7 +50,7 @@ z
 
 @misc{Spielman:2007,
     Author = {{Spielman}, Daniel A.},
-    Title = {{Spectral Graph Theory - Lecture 7}},
+    Title = {Spectral Graph Theory - {L}ecture 7},
     Institution = {Yale University},
     Year = 2007,
     Note = {Available at: \url{https://www.cs.yale.edu/homes/spielman/462/2007/lect7-07.pdf}}
@@ -68,7 +68,7 @@ z
 
 @article{Kelly:2020,
     Author = {{Kelly}, Terence},
-    Title = {{Compressed Sparse Row Format for Representing Graphs}},
+    Title = {Compressed Sparse Row Format for Representing Graphs},
     Journal = {;login: The Magazine of USENIX \& The Advanced Computing Systems Association},
     Volume = {45},
     Number = {4},
@@ -90,14 +90,14 @@ z
 
 @phdthesis{Bai:2007,
     Author = {{Bai}, Xiao},
-    Title = {{Heat Kernel Analysis On Graphs}},
+    Title = {Heat Kernel Analysis On Graphs},
     School = {University of York},
     Year = 2007
 }
 
 @article{Bai:2010,
     Author = {{Bai}, Xiao and {Hancock}, Edwin R.},
-    Title = {{Heat Kernels, Manifolds and Graph Embedding}},
+    Title = {Heat Kernels, Manifolds and Graph Embedding},
     Journal = {Pattern Recognition},
     Note = {University of California, Davis},
     Year = 2010,
@@ -106,14 +106,14 @@ z
 
[…]
