### 2025-09-27 #9002 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
