### 2025-09-27 #9002 @danielskatz
Congratualtions to @Mec-iS (Lorenzo Moriondo) on your JOSS publication!!

And thanks to @sstadick and @DiogoRibeiro7 for reviewing!
JOSS fully depends on volunteers and we couldn't do this without you
